photo-search-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
asyncpg==0.29.0
databases==0.8.0
psycopg2-binary==2.9.9

# MinIO
minio==7.2.0

# Image Processing
Pillow==10.1.0

# ML/DL
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
numpy==1.24.3

# Vector Operations
scikit-learn==1.3.2

# Utilities
aiofiles==23.2.0
httpx==0.25.2